Update German translation for "examiner interests" (ECS-12)
The translations of the phrase "Financing of the study and information on financial or other interests of the examiner" were updated in both locale/de/LC_MESSAGES/django.po and templates/submissions/pdf/section_7.html. This change simplifies the language and makes it more understandable by using the term "potential conflicts of interest".