Add watermark and improve CSS formatting in PDF view (ECS-25)
Added a watermark overlay to the PDF view and improved the presentation of headers within the template by adjusting the text format and orientation. Also added functionality to the Docstash model to render a preview PDF and enable dynamic content updates from multiple formsets. This includes provisions for multicentric studies, investigator information, and other related fields in the submission form.