Restrict preview feature to new submissions only
The preview feature was previously allowed for all submissions. Now, it has been modified to be functional only when a new submission is created. This change includes the setup of a countdown timer for the preview generation cooldown, which is now also exclusive to new submissions. This is reflected in the template and accompanying JavaScript for the submission form.